requirements.txt
openai-agents
aiosqlite
openai-agents[viz]
fastapi
uvicorn[standard]
openai-agents[redis]